Small refactoring and removing old code
